+++ sdk/python/pulumi_gcp/secretmanager/__init__.py
@@ -8,3 +8,4 @@ from .secret_iam_binding import *
 from .secret_iam_member import *
 from .secret_iam_policy import *
 from .secret_version import *
+from .get_secret_version import *